lib_calcFullscreenArgs(): center window properly if taskbar is shown
If the taskbar was shown, then the game window was not centered properly
in all cases.

Instead of centering the window relative to the monitor, it was centered
relative to the screen area that the window is allowed to reside in.

This screen area does not include the MS WIndows taskbar if taskbar is
allowed to be shown, which distorted the centering of the window.

This problem has now been fixed.

BONUS

- Some code refactoring.
- More comments to make code easier to understand.
- Also return the monitor area.