Added taskbar feature: ghost (name pending)
Previously the script could either hide or show the MS Windows taskbar
(system tray).

This choice was made by setting taskbar variable to either "show" or "hide".
(script args for setting this is on my todo list)

A new option for taskbar has been implemented: ghost (name pending)

This option will further limit the allowed screen area for the window to
recide in. Not only will the taskbar area be excluded, but also an area
the same size as the tasbar at the opposite side of the monitor. I
imagine this to be a invisible "ghost" twin of the taskbar (hence the
name).

Not sure if this feature will be used, but now it is here.