Monitor autoselect: check all window corners
When autoselecting a monitor, the monitor that contains the window
should be selected.

Before, only the window top left corner was checked if it exists inside
a monitor area. But a window can theoretically be placed slightly
outside a monitor, leaving the top left corner not visible. This would
make the check fail and the primary monitor would be selected.

This change adds a second check if the first one fails: It selects the
first monitor that overlaps any part of the window.

This way the window should always be fullscreened on a monitor it is
visible on.