Taskbar feature ghost renamed to show3
I have a hard time finding a good name for this feature. The feature is
also very hard to explain.

The visibility for the Windows Taskbar now has four options:
- hide
- show
- show2
- show3

Show, show2, and show3, all will display the taskbar. The difference is
how the game window will be centered on the monitor.

- show: game window is centered relative to the center of the monitor.
- show2: game window is centered relative to the center of the allowed
  window area (monitor area - tasbar area).
- show3: allowed window area is reduced by the taskbar area twice (on
  both side of the monitor).

Not sure if anyone would actually find this feature useful. But know it
is in there.