lib_calcFullscreenArgs(): Can now receive window size'n'position
Function lib_calcFullscreenArgs() could only receive a windows hWnd and
suery its current size and position.

Now it can receive the window size and position as an argument.

I plan to be able to resize the window multiple times during a run. This
means I need to be able to calculate the fullscreen parameters based on
the original window size, not on the current window size.