Always create background: move pixel code
To be able to resize window multiple times during a run, the
background overlay cannot only be created when it is needed. It
needs to always exist and be hidden when it is not needed.

CHANGE 2:
The code for creating the background overlay window is moved earlier in
the code.