Calc fullscreen props AFTER removing window border
Before the script tried to figure out what dimentions the window would
get without borders before the border was removed. This usually works
BUT some game windows refuses to get resized as we want.

Now the window state is fetched AFTER the window border has been
removed.

This led to an overall improvement in code quality and readability (in
my opinion).