Autoselect monitor that has biggest area of window
Changes the algorithm for autoselecting the monitor to fullscreen the
window on.

Before it selected the monitor that contained the window (x,y)
coordinate (top-left corner). But I found out that the visible corner
and the actual window coordinate does not always match. So a window
close to the left side of a monitor may have its coordinate on the next
monitor to the left.

To fix this,m the algorithm was changed to autoselect the monitor that
contains the biggest area of the window. If the window overlaps 2 or
more monitor, the monitor that overlaps the biggest area of the window
will be selected.