BUGFIX: stdlib: Inspect() no work on unknown class
The idea is for function Inspect() to fallback to calling
Object.Inspect() if the class does not have an Inspect() method.

This failed due to a bug, which has been fixed in this commit.

Function IsSet(Var) will check if variable Var contains a value. But the
code was written like IsSet(Var) would check if variable Varname
contained a value (if Var := "VarName"). This made the if-clause always
return true.

EXTRA

Some classes are subclasses, which means Type(subClassObj) would return
a name containing a dot (ex: "Gui.Radio").

Code has now been added that replaces the dot with an underscore.