Refactored code into synchronizeWithWindow()
Moved some global code into threee new functions:

- synchronizeWithWindow()
- CollectWindowInfo(hWnd, &cnfg)
- ConsolePrintWindowInfo(cnfg)

REASON
The code is going to need to call this code multiple times, so it needs
to be in functions.

The idea is for the Console window to have buttons and other controls
that will let users reselect which game window to make fullscreen.