DRAFT: monitor radios: no untoggle, add auto
Before there were only a radio button for each monitor. And clicking on
the selected radio button unselected it (making so no monitor radio
button was selected).

This behavior is not common and could lead to confusion.

This commit changes this so an extra radio button exists that is called
"Auto", which is meant to mean that the monitor will be autoselected.
And this new radio button is now selected by default.

The idea is for all radio button groups to always have one radio button
selected at all times, so no usage confusion occurs.