Arg Parsing: warn on unknown args
Change sounds simple, but actually touched a surpricing amount of code.

The most important edit was to change the return type of parseArgs(args)
from an a list of key-value pairs to a list of key-(value,argStr) pairs.

`argStr` here contains the original argument string from when the script
was started.