restoreWindowState() -> modifyWindowState()
Function restoreWindowState() was only used to change game window
from fullscreen mode to window mode.

By changing it to modifyWindowState() it is now used for all state
modifications on the game window, both for going fullscreen mode and for
going window mode.

This change removed unnecessary duplicated code, and hopefully made the
code easier to understand.