deactivateFullscreen(): Zero out window_mode and fullscreen_mode
Since all properties in both window_mode and fullscreen_mode is modified
everytime fullscreen mode is activated, their old values are not used
again when returning to window mode.

I therefore feel it is a good idea to clean out the old values, so they
are not reused by mistake. Maybe a future update will introduce such a
bug in the code.