BUGFIX: resize=fit broken when game proportions matches monitor
If the game window has the exact same proportions as the monitor, this
would mean that resize=fit would make the game window cover the shole
monitor perfectly during fullscreen.

OR, that should have been. Instead the game window was just made wider.

This commit fixes this simple but confusing bug.