BUGFIX: Always runs WinWait() even if no CLI params
By using CLI params the script can be configured to wait for a specific
game window to appear, and then make it fullscreen.

Even when no CLI params are present the script tries to wait for a game
window with an empty title (empty string "").

This commit fixes the bug by rewriting the affected code.