Disable mainGui when waiting for game window to show up
If CLI args are used to find a game window, the script will run
function WinWait() which completely blocks the main thread until the
game window shows up.

It is therefore probably a bad idea if the user interacted with the
mainGui.

As a bonus, this indicates much clearer to the user that we are waiting
for something to happen.