settings.quit_together: Autoconfigure using temporary global variable
Global setting settings.quit_together is supposed to only be used as a
boolean (true OR false).

But the code temporarily stored a string in it to determine if it should
be autoconfigured to true or false later in the code.

This commit changes so settings.quit_together is always a boolean.