REFACTOR: Globals read data from mainGui, lots of code re-arrangement
I think this is the second really messy commit I have done in this repo.
Sorry.

The main change is that global variables are no longer the center of the
code. As an exaple: The global variables are no longer defined at the
top of the script.

Instead the Gui code is at the top, and the global variables mostly read
and write to the Controls in the Gui.

The rest of the commit is only me re-arranging existing code.

WHY THE CHANGE:

Lots of settings were previously stored both in globals AND in the Gui.
The code for syncing all settings between globals adn Gui were buggy and
hard to get right.

So instead the Gui is now used as the only storage, and the globals
mostly read and write its data to the Gui instead of storing it
themselves.

There are still some data not stored in the Gui that are stored in the
globals.