BUGFIX: On error, still belived game was in fullscreen
When failing to calculate the fullscreen parameters, the game window is
restored to window mode. But the code thought the game window was in
fullsceen mode.

If the user changed settings so the code did not reach a bug, then the
window would be displayed in window mode and have the background overlay
painted around it.