Taskbar: Sometimes failed to go fullscreen on multi-monitor setup
On multi-monitor setups where the window taskbar is configured to show
up on all monitors, the code mistakenly thought it found 2 or more
taskbars on the same monitor.

This led to the code setting an error message, which led to the code
abort the fullscreen procedure (returning to window mode).

Reason for the bug: OF BY 1 ERROR ! ! !